+++ src/Symfony/Component/Translation/CHANGELOG.md
@@ -14,6 +14,7 @@ CHANGELOG
  * Re-add `PoEditorProvider`
  * Add `translation:update-xliff-sources` command to fill the `<source>` tags of XLIFF files with the default locale's translations
  * Add support for ternary and null-coalescing expressions when extracting translation messages from PHP files
+ * Add support for `match` expressions, interpolated strings, variables and values returned by methods when extracting translation messages from PHP files
  * Add argument `$skipValidationDirs` to `XliffFileLoader` to load the XLIFF files of these directories without validating them against the schema
 
 8.1